+++ bindings/nostr-sdk-flutter/rust/src/frb_generated.rs
@@ -43,7 +43,7 @@ flutter_rust_bridge::frb_generated_boilerplate!(
     default_rust_auto_opaque = RustAutoOpaqueMoi,
 );
 pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_VERSION: &str = "2.6.0";
-pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = 440607370;
+pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = 1878779762;
 
 // Section: executor
 
@@ -1145,40 +1145,6 @@ fn wire__crate__api__protocol__key__Keys_secret_key_impl(
         },
     )
 }
-fn wire__crate__api__protocol__key__public_key__PublicKey_from_hex_impl(
-    ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
-    rust_vec_len_: i32,
-    data_len_: i32,
-) -> flutter_rust_bridge::for_generated::WireSyncRust2DartSse {
-    FLUTTER_RUST_BRIDGE_HANDLER.wrap_sync::<flutter_rust_bridge::for_generated::SseCodec, _>(
-        flutter_rust_bridge::for_generated::TaskInfo {
-            debug_name: "PublicKey_from_hex",
-            port: None,
-            mode: flutter_rust_bridge::for_generated::FfiCallMode::Sync,
-        },
-        move || {
-            let message = unsafe {
-                flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(
-                    ptr_,
-                    rust_vec_len_,
-                    data_len_,
-                )
-            };
-            let mut deserializer =
-                flutter_rust_bridge::for_generated::SseDeserializer::new(message);
-            let api_public_key = <String>::sse_decode(&mut deserializer);
-            deserializer.end();
-            transform_result_sse::<_, flutter_rust_bridge::for_generated::anyhow::Error>(
-                (move || {
-                    let output_ok = crate::api::protocol::key::public_key::_PublicKey::from_hex(
-                        &api_public_key,
-                    )?;
-                    Ok(output_ok)
-                })(),
-            )
-        },
-    )
-}
 fn wire__crate__api__protocol__key__public_key__PublicKey_from_slice_impl(
     ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
     rust_vec_len_: i32,
@@ -1389,40 +1355,6 @@ fn wire__crate__api__protocol__key__public_key__PublicKey_to_nostr_uri_impl(
                         crate::api::protocol::key::public_key::_PublicKey::to_nostr_uri(
                             &*api_that_guard,
                         )?;
-                    Ok(output_ok)
-                })(),
-            )
-        },
-    )
-}
-fn wire__crate__api__protocol__key__secret_key__SecretKey_from_hex_impl(
-    ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
-    rust_vec_len_: i32,
-    data_len_: i32,
-) -> flutter_rust_bridge::for_generated::WireSyncRust2DartSse {
-    FLUTTER_RUST_BRIDGE_HANDLER.wrap_sync::<flutter_rust_bridge::for_generated::SseCodec, _>(
-        flutter_rust_bridge::for_generated::TaskInfo {
-            debug_name: "SecretKey_from_hex",
-            port: None,
-            mode: flutter_rust_bridge::for_generated::FfiCallMode::Sync,
-        },
-        move || {
-            let message = unsafe {
-                flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(
-                    ptr_,
-                    rust_vec_len_,
-                    data_len_,
-                )
-            };
-            let mut deserializer =
-                flutter_rust_bridge::for_generated::SseDeserializer::new(message);
-            let api_secret_key = <String>::sse_decode(&mut deserializer);
-            deserializer.end();
-            transform_result_sse::<_, flutter_rust_bridge::for_generated::anyhow::Error>(
-                (move || {
-                    let output_ok = crate::api::protocol::key::secret_key::_SecretKey::from_hex(
-                        &api_secret_key,
-                    )?;
                     Ok(output_ok)
                 })(),
             )
@@ -2270,7 +2202,7 @@ fn pde_ffi_dispatcher_primary_impl(
         1 => wire__crate__api__client__Client_add_relay_impl(port, ptr, rust_vec_len, data_len),
         2 => wire__crate__api__client__Client_connect_impl(port, ptr, rust_vec_len, data_len),
         4 => wire__crate__api__client__Client_send_event_impl(port, ptr, rust_vec_len, data_len),
-        45 => wire__crate__api__init_app_impl(port, ptr, rust_vec_len, data_len),
+        43 => wire__crate__api__init_app_impl(port, ptr, rust_vec_len, data_len),
         _ => unreachable!(),
     }
 }
@@ -2314,80 +2246,70 @@ fn pde_ffi_dispatcher_sync_impl(
         22 => wire__crate__api__protocol__key__Keys_parse_impl(ptr, rust_vec_len, data_len),
         23 => wire__crate__api__protocol__key__Keys_public_key_impl(ptr, rust_vec_len, data_len),
         24 => wire__crate__api__protocol__key__Keys_secret_key_impl(ptr, rust_vec_len, data_len),
-        25 => wire__crate__api__protocol__key__public_key__PublicKey_from_hex_impl(
+        25 => wire__crate__api__protocol__key__public_key__PublicKey_from_slice_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        26 => wire__crate__api__protocol__key__public_key__PublicKey_from_slice_impl(
+        26 => wire__crate__api__protocol__key__public_key__PublicKey_parse_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        27 => wire__crate__api__protocol__key__public_key__PublicKey_parse_impl(
+        27 => wire__crate__api__protocol__key__public_key__PublicKey_to_bech32_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        28 => wire__crate__api__protocol__key__public_key__PublicKey_to_bech32_impl(
+        28 => wire__crate__api__protocol__key__public_key__PublicKey_to_hex_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        29 => wire__crate__api__protocol__key__public_key__PublicKey_to_hex_impl(
+        29 => wire__crate__api__protocol__key__public_key__PublicKey_to_nostr_uri_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        30 => wire__crate__api__protocol__key__public_key__PublicKey_to_nostr_uri_impl(
+        30 => wire__crate__api__protocol__key__secret_key__SecretKey_from_slice_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        31 => wire__crate__api__protocol__key__secret_key__SecretKey_from_hex_impl(
+        31 => wire__crate__api__protocol__key__secret_key__SecretKey_generate_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        32 => wire__crate__api__protocol__key__secret_key__SecretKey_from_slice_impl(
+        32 => wire__crate__api__protocol__key__secret_key__SecretKey_parse_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        33 => wire__crate__api__protocol__key__secret_key__SecretKey_generate_impl(
+        33 => wire__crate__api__protocol__key__secret_key__SecretKey_to_bech32_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        34 => wire__crate__api__protocol__key__secret_key__SecretKey_parse_impl(
+        34 => wire__crate__api__protocol__key__secret_key__SecretKey_to_secret_hex_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        35 => wire__crate__api__protocol__key__secret_key__SecretKey_to_bech32_impl(
+        35 => wire__crate__api__protocol__event__tag__Tag_as_vec_impl(ptr, rust_vec_len, data_len),
+        36 => wire__crate__api__protocol__event__tag__Tag_content_impl(ptr, rust_vec_len, data_len),
+        37 => wire__crate__api__protocol__event__tag__Tag_is_protected_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        36 => wire__crate__api__protocol__key__secret_key__SecretKey_to_secret_hex_impl(
-            ptr,
-            rust_vec_len,
-            data_len,
-        ),
-        37 => wire__crate__api__protocol__event__tag__Tag_as_vec_impl(ptr, rust_vec_len, data_len),
-        38 => wire__crate__api__protocol__event__tag__Tag_content_impl(ptr, rust_vec_len, data_len),
-        39 => wire__crate__api__protocol__event__tag__Tag_is_protected_impl(
-            ptr,
-            rust_vec_len,
-            data_len,
-        ),
-        40 => {
+        38 => {
             wire__crate__api__protocol__event__tag__Tag_is_reply_impl(ptr, rust_vec_len, data_len)
         }
-        41 => wire__crate__api__protocol__event__tag__Tag_is_root_impl(ptr, rust_vec_len, data_len),
-        42 => wire__crate__api__protocol__event__tag__Tag_kind_impl(ptr, rust_vec_len, data_len),
-        43 => wire__crate__api__protocol__event__tag__Tag_parse_impl(ptr, rust_vec_len, data_len),
-        44 => wire__crate__api__protocol__event__tag__Tag_to_vec_impl(ptr, rust_vec_len, data_len),
+        39 => wire__crate__api__protocol__event__tag__Tag_is_root_impl(ptr, rust_vec_len, data_len),
+        40 => wire__crate__api__protocol__event__tag__Tag_kind_impl(ptr, rust_vec_len, data_len),
+        41 => wire__crate__api__protocol__event__tag__Tag_parse_impl(ptr, rust_vec_len, data_len),
+        42 => wire__crate__api__protocol__event__tag__Tag_to_vec_impl(ptr, rust_vec_len, data_len),
         _ => unreachable!(),
     }
 }