+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * nostr: update `FromBech32::from_bech32` method signature ([Yuki Kishimoto])
 * nostr: update `NostrURI::from_nostr_uri` method signature ([Yuki Kishimoto])
 * nostr: remove generics from parsing methods in `key` module ([Yuki Kishimoto])
+* bindings: remove redundant parsing methods from `EventId`, `Coordinate`, `PublicKey` and `SecretKey` ([Yuki Kishimoto])
 
 ### Changed